src/RandomForestUnitTest.java
import org.junit.jupiter.api.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class RandomForestUnitTest {

    /**
     *  This is a test class for the RandomForest class.
     *  It aims to test all the public methods declared in the RandomForest class.
     */

    @Test
    public void testRandomForestConstruction() {
        // Testing constructor and initialization
        RandomForest forest = new RandomForest(10, 5, 3);
        assertNotNull(forest);
    }

    @Test
    public void testRandomForestTrain() {
        // Testing train method
        RandomForest forest = new RandomForest(10, 5, 3);
        List<ClimbingFrame> data = new ArrayList<>();
        assertDoesNotThrow(() -> forest.train(data));
    }

    @Test
    public void testRandomForestPredict() {
        // Testing predict method
        RandomForest forest = new RandomForest(10, 5, 3);
        List<ClimbingFrame> data = new ArrayList<>();
        forest.train(data);
        ClimbingFrame frame = new ClimbingFrame(); // assuming ClimbingFrame has a default constructor
        int prediction = forest.predict(frame);
        assertNotNull(prediction);
    }

    @Test
    public void testRandomForestPredictFromString() {
        // Testing predict from string method
        RandomForest forest = new RandomForest(10, 5, 3);
        List<ClimbingFrame> data = new ArrayList<>();
        forest.train(data);
        int prediction = forest.predictFromString("test string"); // replace "test string" with a valid string representation
        assertNotNull(prediction);
    }

    @Test
    public void testSaveLoadModel() throws Exception {
        // Testing the save and load model methods
        RandomForest forest1 = new RandomForest(10, 5, 3);
        List<ClimbingFrame> data = new ArrayList<>();
        forest1.train(data);
        String filePath = "test_model.dat";
        forest1.saveModel(filePath);
        RandomForest forest2 = RandomForest.loadModel(filePath);
        assertEquals(forest1.numTrees, forest2.numTrees);
        assertEquals(forest1.maxDepth, forest2.maxDepth);
        assertEquals(forest1.maxFeatures, forest2.maxFeatures);
    }
}